+++ tests/operators/test_hex_operator.py
@@ -17,12 +17,17 @@ def test_my_custom_operator_execute_no_trigger(dag, requests_mock):
         json={"projectId": "ABC-123", "runId": "1"},
     )
 
-    mock_status = {"projectId": "abc-123", "status": "COMPLETED"}
+    mock_status = {
+        "projectId": "ABC-123",
+        "status": "COMPLETED",
+        "runUrl": "https://example.com/run/1",
+    }
     requests_mock.get(
         "https://www.httpbin.org/api/v1/project/abc-123/run/1",
         headers={"Content-Type": "application/json"},
         json=mock_status,
     )
+
     dagrun = dag.create_dagrun(
         state=DagRunState.RUNNING,
         execution_date=DATA_INTERVAL_START,
@@ -33,6 +38,7 @@ def test_my_custom_operator_execute_no_trigger(dag, requests_mock):
     ti = dagrun.get_task_instance(task_id=TEST_TASK_ID)
     ti.task = dag.get_task(task_id=TEST_TASK_ID)
     ti.run(ignore_ti_state=True)
+
     assert ti.state == TaskInstanceState.SUCCESS
     json = requests_mock.request_history[0].json()
     assert json["inputParams"]["input_date"][0:4] == str(DATA_INTERVAL_START.year)